pyproject.toml
[tool.poetry]
name = "locust-pinecone"
version = "0.1.0"
description = "A load-testing tool for Pinecone, built on the Locust framework"
authors = ["Pinecone Systems, Inc. <support@pinecone.io>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
locust = "^2.24.1"
pandas = "^2.2.1"
google-cloud-storage = "^2.14.0"
grpcio = "^1.62.0"
python-dotenv = "^1.0.1"
pyarrow = "^15.0.0"
pinecone-client = {version = "^3.2.2", extras = ["grpc"]}
tabulate = "^0.9.0"
numpy = "^1.26.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
flake8 = "^7.0.0"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
